requirements.txt
anyio==3.6.2
click==8.1.3
colorama==0.4.6
envyaml==1.10.211231
fastapi==0.95.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
python-jose==3.3.0
pydantic==1.10.7
pymssql==2.2.7
PyYAML==6.0
sniffio==1.3.0
SQLAlchemy==2.0.9
starlette==0.26.1
typing_extensions==4.5.0
uvicorn==0.21.1
passlib~=1.7.4
PyMySQL~=1.0.3
requests~=2.31.0